Work on refining DP introduction
This is something of a difficult concept to succinctly describe.  The
original text here was pretty good, but there were a few inaccuracies
(the one where \epsilon and sensitivity were confused in particular).

This attempts to restructure text a little, adjusting for flow.  I moved
a couple of paragraphs around in that process.  For instance, the note
about robustness against malicious inputs really belongs in the section
on sensitivty capping.