Encrypted match keys aren't portable (#18)
* Encrypted match keys aren't portable

I didn't note that the encrypted match key is bound to its intended
type.  I don't think that we care about that in the interface to the
browser.  We are relying on the helpers to enforce limits based on how
events are labeled, but we don't need to distinguish types at the time
they are generated.

* Update IPA-End-to-End.md

Co-authored-by: Martin Thomson <mt@lowentropy.net>

Co-authored-by: Erik Taubeneck <erik.taubeneck@gmail.com>