Add some thoughts on the application of encryption
This contains some fairly concrete ideas about how match keys are
protected.

It also outlines how we might use TLS.

The most interesting part is the query submission process, which I
expect will need some more discussion.

Until now, we'd more or less assumed a model closer to that in PPM.
There, all data is submitted to a leader, which redistributes the data.

But that assumes a very specific submission model that doesn't really
match the one we've settled on.  In looking at this closer, I've
basically concluded that the complications involved in interleaving
multiple records into a single query, with separate streams of
encryption, are significant enough that a different approach is
warranted.

I'm advocating for the record collector submitting data to each of the
helper parties.  I explore an option where the three flows all go via
a single leader, but I think that just adds inefficiency.

Feedback is very much welcome, as this is still quite drafty.  I've
opened issues to track some of the things that came up in this process.