Include match key provider origin
This now needs to be made available to the helpers in order to decrypt.
This allows the helpers to enforce consistency of choice of match key
provider.

Also switch to registrable domain for the site identity.  This is a
smaller string than origin, but much more challenging to compute.
The definition of registrable domain changes over time as the public
suffix list changes.  This is not expected to be a problem as the
browser can report the value it used, which can be saved alongside the
encrypted match key.

The real challenge there is in how helpers treat queries where a change
has occurred.  For the most part, this should not be a problem for
source events in trigger queries or trigger events in source queries as
similar sites can just be assigned appropriate breakdown keys.  For
other events, the set of sites from both before and after a change
should be counted toward a single privacy budget.  If helpers become
aware of a change after queries are made (that is, helpers are slower
than browsers to pick up a change), that is a problem.

For example, x.example is added to the PSL and now a.x.example and
b.x.example are considered to be different sites.  Any query that
includes events from either side of the change (plus adjustments to
allow for the change to propagate) would count a.x.example, b.x.example,
*.x.example, and x.example toward the same budget.