A frank appraisal of user benefit
A common question that comes up in this context is what benefits users
of the web might expect to see as a result.

We've spoken about this in PATCG meetings on several occasions.
Supporting advertising has no direct user benefit.
Indirect benefits exist however.
This touches the more prominent reasons.

Also, this acknowledges the bad motivation that is the hope that
tracking will cease just because attribution is possible.  That's a
fanciful idea, but not a realistic one.  What is possible (and what I
didn't include because it gets tricky) is that with attribution being
effectively possible, it might become easier to more strictly limit
cross-site communication.