It got worse
After we discussed this, I went back and did some more thinking.

Initially, it seemed pretty simple: each app is its own browser.  For
the most part, this works perfectly.

But then I realized that the app is in a position to misrepresent site
identity.  That's not something I have an answer for.